Rename 'data' as 'terminal-data'.
If you want to use the socket connection for transporting anything other
than terminal data, then the name 'data' is probably too generic. If we
make it more descriptive, then the sources and destinations are clearer.

Use case: I want to have a terminal window editing a file, but also the
full file contents below. And, to transmit those contents, I want to
re-use the socket connection. It would be helpful if the naming of the
terminal data were descriptive.

App.js also now exports the the socket (io).